requirements.txt
psycopg2-binary==2.8.6
django==2.2
djangorestframework==3.9.2
djangorestframework-gis==0.14
drf-nested-routers==0.91
django_extensions==2.1.6
django-cors-headers==2.5.2
django-cors-middleware==1.3.1
django_filter==2.1.0
boto3==1.9.126
simpleflake==0.1.5
unicodecsv==0.14.1
gunicorn==19.9.0
django-autofixture==0.12.1
coreapi==2.3.3
pyyaml==3.13
django-rest-swagger==2.2.0
django-rest-auth==0.9.5
